+++ distributions/univariate/continuous/DegenerateRand.cpp
@@ -35,12 +35,12 @@ double DegenerateRand::Variate() const
     return a;
 }
 
-double DegenerateRand::Mean() const
+long double DegenerateRand::Mean() const
 {
     return a;
 }
 
-double DegenerateRand::Variance() const
+long double DegenerateRand::Variance() const
 {
     return 0.0;
 }
@@ -72,12 +72,12 @@ double DegenerateRand::Mode() const
     return a;
 }
 
-double DegenerateRand::Skewness() const
+long double DegenerateRand::Skewness() const
 {
     return NAN;
 }
 
-double DegenerateRand::ExcessKurtosis() const
+long double DegenerateRand::ExcessKurtosis() const
 {
     return NAN;
 }
