+++ distributions/univariate/continuous/LogisticRand.h
@@ -38,12 +38,12 @@ public:
     double S(const double & x) const override;
     double Variate() const override;
 
-    double Mean() const override;
-    double Variance() const override;
+    long double Mean() const override;
+    long double Variance() const override;
     double Median() const override;
     double Mode() const override;
-    double Skewness() const override;
-    double ExcessKurtosis() const override;
+    long double Skewness() const override;
+    long double ExcessKurtosis() const override;
 private:
     double quantileImpl(double p) const override;
     double quantileImpl1m(double p) const override;
