+++ distributions/univariate/continuous/MarchenkoPasturRand.h
@@ -76,11 +76,11 @@ private:
     double Moment(int n) const;
 
 public:
-    double Mean() const override;
-    double Variance() const override;
+    long double Mean() const override;
+    long double Variance() const override;
     double Mode() const override;
-    double Skewness() const override;
-    double ExcessKurtosis() const override;
+    long double Skewness() const override;
+    long double ExcessKurtosis() const override;
 
 private:
     double quantileImpl(double p) const override;
