+++ distributions/univariate/continuous/NormalRand.h
@@ -91,8 +91,8 @@ private:
 
 public:
     double Moment(int n) const;
-    double ThirdMoment() const override { return Moment(3); }
-    double FourthMoment() const override { return Moment(4); }
+    long double ThirdMoment() const override { return Moment(3); }
+    long double FourthMoment() const override { return Moment(4); }
 
     /**
      * @fn FitLocation
